requirements.txt
tkinter
pyserial
numpy
matplotlib